+++ src/dock_state/mod.rs
@@ -12,6 +12,7 @@ pub mod window_state;
 
 pub use surface::Surface;
 pub use surface_index::SurfaceIndex;
+use tree::node::LeafNode;
 pub use window_state::WindowState;
 
 use egui::Rect;
@@ -186,8 +187,8 @@ impl<Tab> DockState<Tab> {
         &mut self,
         (surface_index, node_index, tab_index): (SurfaceIndex, NodeIndex, TabIndex),
     ) {
-        if let Some(Node::Leaf { active, .. }) = self[surface_index].nodes.get_mut(node_index.0) {
-            *active = tab_index;
+        if let Some(Node::Leaf(leaf)) = self[surface_index].nodes.get_mut(node_index.0) {
+            leaf.active = tab_index;
         }
     }
 
@@ -403,7 +404,7 @@ impl<Tab> DockState<Tab> {
             .flat_map(|(surface_index, surface)| {
                 surface
                     .iter_nodes()
-                    .map(move |node| (SurfaceIndex(surface_index), node))
+                    .map(move |node| (surface_index.into(), node))
             })
     }
 
@@ -415,7 +416,7 @@ impl<Tab> DockState<Tab> {
             .flat_map(|(surface_index, surface)| {
                 surface
                     .iter_nodes_mut()
-                    .map(move |node| (SurfaceIndex(surface_index), node))
+                    .map(move |node| (surface_index.into(), node))
             })
     }
 
@@ -427,7 +428,7 @@ impl<Tab> DockState<Tab> {
             .flat_map(|(surface_index, surface)| {
                 surface
                     .iter_all_tabs()
-                    .map(move |(node_index, tab)| ((SurfaceIndex(surface_index), node_index), tab))
+                    .map(move |(node_index, tab)| ((surface_index.into(), node_index), tab))
             })
     }
 
@@ -441,7 +442,7 @@ impl<Tab> DockState<Tab> {
             .flat_map(|(surface_index, surface)| {
                 surface
                     .iter_all_tabs_mut()
-                    .map(move |(node_index, tab)| ((SurfaceIndex(surface_index), node_index), tab))
+                    .map(move |(node_index, tab)| ((surface_index.into(), node_index), tab))
             })
     }
 
@@ -464,6 +465,18 @@ impl<Tab> DockState<Tab> {
             .iter()
             .filter_map(|surface| surface.node_tree())
             .flat_map(|nodes| nodes.iter())
+    }
+
+    /// Returns an immutable [`Iterator`] of all [``LeafNode``]s in the dock state.
+    pub fn iter_leaves(&self) -> impl Iterator<Item = (SurfaceIndex, &LeafNode<Tab>)> {
+        self.iter_all_nodes()
+            .filter_map(|(index, node)| node.get_leaf().map(|leaf| (index, leaf)))
+    }
+
+    /// Returns a mutable [`Iterator`] of all [``LeafNode``]s in the dock state.
+    pub fn iter_leaves_mut(&mut self) -> impl Iterator<Item = (SurfaceIndex, &mut LeafNode<Tab>)> {
+        self.iter_all_nodes_mut()
+            .filter_map(|(index, node)| node.get_leaf_mut().map(|leaf| (index, leaf)))
     }
 
     /// Returns a new [`DockState`] while mapping and filtering the tab type.
@@ -556,6 +569,28 @@ impl<Tab> DockState<Tab> {
             !surface.is_empty()
         });
     }
+
+    /// Find a tab based on the conditions of a functino.
+    ///
+    /// Returns in which node and where in that node the tab is.
+    ///
+    /// The returned [`NodeIndex`] will always point to a [`Node::Leaf`].
+    ///
+    /// In case there are several hits, only the first is returned.
+    pub fn find_tab_from(
+        &self,
+        predicate: impl Fn(&Tab) -> bool,
+    ) -> Option<(SurfaceIndex, NodeIndex, TabIndex)> {
+        for &surface_index in self.valid_surface_indices().iter() {
+            if self.surfaces[surface_index.0].is_empty() {
+                continue;
+            }
+            if let Some((node_index, tab_index)) = self[surface_index].find_tab_from(&predicate) {
+                return Some((surface_index, node_index, tab_index));
+            }
+        }
+        None
+    }
 }
 
 impl<Tab> DockState<Tab>
@@ -572,14 +607,7 @@ where
     ///
     /// See also: [`find_main_surface_tab`](DockState::find_main_surface_tab)
     pub fn find_tab(&self, needle_tab: &Tab) -> Option<(SurfaceIndex, NodeIndex, TabIndex)> {
-        for &surface_index in self.valid_surface_indices().iter() {
-            if !self.surfaces[surface_index.0].is_empty() {
-                if let Some((node_index, tab_index)) = self[surface_index].find_tab(needle_tab) {
-                    return Some((surface_index, node_index, tab_index));
-                }
-            }
-        }
-        None
+        self.find_tab_from(|tab| tab == needle_tab)
     }
 
     /// Find the given tab on the main surface.